# Improve rendering performance