--- a/src/Markdig/Polyfills/NullableAttributes.cs
+++ b/src/Markdig/Polyfills/NullableAttributes.cs
@@ -15,6 +15,9 @@ namespace System.Diagnostics.CodeAnalysis
 
         public bool ReturnValue { get; }
     }
+
+    [AttributeUsage(AttributeTargets.Property | AttributeTargets.Field | AttributeTargets.Parameter, Inherited = false)]
+    public sealed class AllowNullAttribute : Attribute { }
 #endif
 
 #if !NET5_0_OR_GREATER