--- a/src/Markdig/Renderers/Html/HeadingRenderer.cs
+++ b/src/Markdig/Renderers/Html/HeadingRenderer.cs
@@ -2,7 +2,6 @@
 // This file is licensed under the BSD-Clause 2 license. 
 // See the license.txt file in the project root for more information.
 
-using System.Globalization;
 using Markdig.Syntax;
 
 namespace Markdig.Renderers.Html
@@ -25,20 +24,26 @@ namespace Markdig.Renderers.Html
         protected override void Write(HtmlRenderer renderer, HeadingBlock obj)
         {
             int index = obj.Level - 1;
-            string headingText = ((uint)index < (uint)HeadingTexts.Length)
-                ? HeadingTexts[index]
-                : "h" + obj.Level.ToString(CultureInfo.InvariantCulture);
+            string[] headings = HeadingTexts;
+            string headingText = ((uint)index < (uint)headings.Length)
+                ? headings[index]
+                : $"h{obj.Level}";
 
             if (renderer.EnableHtmlForBlock)
             {
-                renderer.Write("<").Write(headingText).WriteAttributes(obj).Write('>');
+                renderer.Write('<');
+                renderer.WriteRaw(headingText);
+                renderer.WriteAttributes(obj);
+                renderer.WriteRaw('>');
             }
 
             renderer.WriteLeafInline(obj);
 
             if (renderer.EnableHtmlForBlock)
             {
-                renderer.Write("</").Write(headingText).WriteLine(">");
+                renderer.Write("</");
+                renderer.WriteRaw(headingText);
+                renderer.WriteLine('>');
             }
 
             renderer.EnsureLine();