… src/Markdig/Extensions/AutoLinks/NormalizeAutoLinkRenderer.cs
… src/Markdig/Renderers/Html/Inlines/EmphasisInlineRenderer.cs
… src/Markdig/Renderers/Html/Inlines/LinkInlineRenderer.cs
… src/Markdig/Renderers/Html/ListRenderer.cs
… src/Markdig/Renderers/Html/QuoteBlockRenderer.cs
… src/Markdig/Renderers/Html/ThematicBreakRenderer.cs
… src/Markdig/Renderers/HtmlRenderer.cs
… src/Markdig/Renderers/IMarkdownObjectRenderer.cs
… src/Markdig/Renderers/MarkdownObjectRenderer.cs
… src/Markdig/Renderers/Normalize/Inlines/LinkInlineRenderer.cs
… src/Markdig/Renderers/RendererBase.cs