Closes #547

API change as mentioned in #547:
```diff
public interface IMarkdownObjectRenderer
{
-   bool Accept(RendererBase renderer, MarkdownObject obj);
+   bool Accept(RendererBase renderer, Type objectType);
}
```


Before
|           Method |     SourceText |      Mean |    Error |   StdDev |    Median |   Gen 0 | Allocated |
|----------------- |--------------- |----------:|---------:|---------:|----------:|--------:|----------:|
| ToHtmlRenderOnly |  MarkdigReadme | 127.31 us | 0.296 us | 0.433 us | 127.36 us |  8.5449 |     36 KB |
| ToHtmlRenderOnly | TracingArticle |  19.92 us | 0.156 us | 0.209 us |  20.04 us |  1.3428 |      6 KB |

After
|           Method |     SourceText |       Mean |     Error |    StdDev |   Gen 0 | Allocated |
|----------------- |--------------- |-----------:|----------:|----------:|--------:|----------:|
| ToHtmlRenderOnly |  MarkdigReadme |  69.793 us | 0.0402 us | 0.0550 us |  8.6670 |     36 KB |
| ToHtmlRenderOnly | TracingArticle |   8.990 us | 0.0297 us | 0.0426 us |  1.3428 |      6 KB |